requirements.txt
streamlit>=1.32.0
yfinance>=0.2.38
curl_cffi>=0.6.2
beautifulsoup4>=4.12.0
plotly>=5.20.0
pandas>=2.0.0
gspread>=5.12.0
google-auth>=2.27.0
